renames nav_bar file
