+++ src/scripts/navBar.js
